pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orthrus"
version = "0.2.0"
description = "Orthrus — ML Data Capture System for Hermes Agent"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "zestehl"},
]
keywords = ["orthrus", "hermes", "agent", "ml", "data-capture", "training-data", "trajectories"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "pyyaml==6.0.2",
    "pydantic==2.11.5",
    "pydantic-settings==2.9.1",
    "structlog==25.3.0",
    "rich==14.0.0",
    "typer==0.24.1",
    "psutil==7.0.0",
    "platformdirs==4.2.0",
    "pyarrow==19.0.1",
    "zstandard==0.23.0",
    "sentence-transformers==4.1.0",
]

# Hermes agent is provided by the host environment or installed separately
# See: https://github.com/NousResearch/hermes-agent

[project.optional-dependencies]
dev = [
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.1.0",
    "ruff==0.15.10",
    "mypy==1.20.0",
    "types-PyYAML==6.0.12.20240311",
    "respx==0.22.0",
    "freezegun==1.5.1",
    "faker==37.4.0",
    "pre-commit==4.2.0",
]

[project.scripts]
orthrus = "orthrus.cli:main_cli"
orthrus-setup = "orthrus.setup:main"
orthrus-audit = "orthrus.orthrus_audit:main"

[project.entry-points."hermes.plugins"]
orthrus = "orthrus.integration:HermesPlugin"

[tool.hatch.build.targets.wheel]
packages = ["src/orthrus"]

# === UV CONFIGURATION ===
[tool.uv]
python-preference = "only-system"

[tool.uv.pip]
annotation-style = "line"

# === DEPENDENCY GROUPS (PEP 735) ===
[dependency-groups]
dev = [
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.1.0",
    "ruff==0.15.10",
    "mypy==1.20.0",
    "types-PyYAML==6.0.12.20240311",
    "respx==0.22.0",
    "freezegun==1.5.1",
    "faker==37.4.0",
    "pre-commit==4.2.0",
]

# === RUFF CONFIGURATION ===
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "N",
    "UP",
    "B",
    "C4",
    "SIM",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

# === MYPY CONFIGURATION ===
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
ignore_missing_imports = true

# NOTE: tests/ is excluded from mypy strict checking — pytest handles type
# verification separately. Run `uv run mypy src/` for strict type coverage.

# === PYTEST CONFIGURATION ===
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

# === COVERAGE CONFIGURATION ===
[tool.coverage.run]
source = ["src/orthrus"]
branch = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if __name__ == .__main__.:"]
